Moved the duplicate functionality from appdataController to duplicateController
