+++ controllers/appdataController.js
@@ -1,4 +1,5 @@
 const { ObjectId } = require("mongodb"); // Import ObjectId
+const { handleDuplicateLead } = require('./duplicateController');
 const config = require("../config/config.js");
 
 const getAppDataCollection = (req) => {
@@ -162,7 +163,7 @@ const createAppData = async (req, res) => {
   try {
     const collection = await getAppDataCollection(req);
     const newData = req.body;
-
+    newData.pageName = "enquiry"; // Set pageName to "enquiry" by default
     console.log("New data to be inserted:", newData);
 
     const insertResult = await collection.insertOne(newData);
@@ -171,27 +172,8 @@ const createAppData = async (req, res) => {
       console.log("New data inserted successfully:", insertResult);
 
       // Check if pageName is "enquiry"
-      if (newData.pageName === "enquiry") {
-        console.log("Page name is enquiry, checking for existing mobile_phone");
-
-        // Check if mobile_phone value is already present in the collection, excluding the newly inserted document
-        console.log("Checking for mobile_phone:", newData.mobile_phone);
-        const existingData = await collection.findOne({
-          mobile_phone: newData.mobile_phone,
-          _id: { $ne: insertResult.insertedId } // Exclude the newly inserted document
-        });
-
-        console.log("Existing data with same mobile_phone:", existingData);
-
-        if (!existingData) {
-          // Create a replica data with pageName changed to "leads"
-          const replicaData = { ...newData, pageName: "leads" };
-          delete replicaData._id; // Remove the _id field to avoid duplicate key error
-          await collection.insertOne(replicaData);
-          console.log('New Lead inserted');
-        } else {
-          console.log('This is a duplicated lead');
-        }
+      if (newData) {
+        await handleDuplicateLead(collection, newData, insertResult.insertedId);
       }
 
       res.status(201).json({
